changed all float to double

+++ frm.c
@@ -1,10 +1,10 @@
 #include "frm.h"
 
 double exp_dist(double a, double l){
-	return -logf(1-a)/l; 
+	return -log(1-a)/l; 
 }
 
-float hyperexp2(float a, float p, float q, float l1, float l2){
+double hyperexp2(double a, double p, double q, double l1, double l2){
 	return exp_dist(a, (p < q) ? l1 : l2);
 }
